修改 IBObject 为 Item ，Item 配置合并到 Spring PropertySource，修复部分问题
1.修改 IBObject 为 Item
==>作为容器 Container 中的所有元素(项)的顶层接口
==>相应修改 BaseObject 、initObject 等命名

2.Item 配置合并到 Spring PropertySource
==>移除原有逐个配置文件，支持合并到一个或多个文件进行配置
    ==>新增 YamlProperySourceFactory 支持 PropertySource 加载 yml 文件
    ==>通过定义不同 name 的 PropertySource 可以实现多个文件中配置的合并和覆盖
==>相应移除 Config 类，移除 saveConfigFile 配置
==>移除 ConfigProcessor，增加 IBConfig 用于属性注入
    ==>解决 ConfigPropcessor 不能方便选择构造方法的问题
    ==>通过 @Configurable 接口标注一个属性或 set 方法用于属性注入
        ==>移除不再需要的 @ProtoConfigurable 注解
    ==>新增 ConfigApplier 接口实现从 IBConifg 注入属性，屏蔽与 IBConfig 交互的细节
        ==>提供 ItemConfigAppiler 方便 Item 的属性注入和 x、y、centerX、centerY 设置
==>临时去除 Control 的部分配置化
    ==>width 、height 等属性默认为 0，由用户自行代码设置，不提供配置注入

3.其他
==>修复部分问题
    ==>修复 BasePhysicsPlugin.checkCollisionBetween collided 未初始化时可能的空指针
    ==>修复 BasePhysicsPlugin.doMove 判断 Property 错误导致碰撞检测不精确
        ==>r.hasProperty(Collider.class) -> Rigidbody.class
==>完善Rect、BasePhysicsPlugin 等单元测试
==>修改 Point x、y 类型为 double，支持精度更高的碰撞检测
    ==>相应修改相关单元测试

+++ ib-core/src/main/java/cn/milai/ib/geometry/Vector.java
@@ -6,8 +6,8 @@ package cn.milai.ib.geometry;
  * @date 2020.03.11
  */
 public class Vector {
-	private long x;
-	private long y;
+	private double x;
+	private double y;
 
 	/**
 	 * 创建从 p1 指向 p2 的向量
@@ -24,7 +24,7 @@ public class Vector {
 	 * @param v
 	 * @return
 	 */
-	public long product(Vector v) {
+	public double product(Vector v) {
 		return x * v.y - y * v.x;
 	}
 }
